requirements.txt
Pillow
newspaper3k
urllib3
streamlit
beautifulsoup4
protobuf==3.20.0
jinja2>=3.0
newspaper3k==0.2.8
nltk==3.6.5
Pillow==9.0.1
requests==2.26.0
streamlit==1.5.0
streamlit-autorefresh==0.2.0
sumy==0.8.1